Rename diagnostic_context to sema_context
Summary: The naming of `diagnostic_context` is quite misleading because diagnostic is not its salient feature. It is also often confused with `diagnostics_engine` resulting in misuses (now mostly fixed), particularly in earlier stages of translation. And some of its diagnostic methods are broken outside of visitation in sema. What it is, is an AST visitor context for semantic analysis that includes the usual AST context state, diagnostic reporting and node metadata cache. Rename it to `sema_context` to make it clear and add a comment.

Reviewed By: yoney

Differential Revision: D61797752

fbshipit-source-id: 926fd5e42d54efa44e99c0fb58174e7dd61a2890