Add list of transitive include schemas to constants file
Summary:
When `--inject-schema` is enabled, injects a special constant that provides references to the schemas of all transitive dependencies of the given program. This allows computation of the full schema of that program without duplicating the information in the binary.
The schema consts are keyed by path hash, which in principle uniquely identifies each Thrift file. Unfortunately there is a legacy include resolution behavior that allows passing paths relative to the main file, which results in the same include having multiple different paths in the AST. In these cases we will be unable to refer to the schema const for the affected file, but to ensure the constants target still compiles we include a weak symbol for each schema const we expect to see returning an empty schema.

Reviewed By: praihan

Differential Revision: D61240139

fbshipit-source-id: f776ddf88e4c3d86082ffb070a09e95b89feac94