Split off partial-lookup from identifiers
Summary:
We already had some special casing of allowing `/` to be a separator for partial paths.

This falls apart because there are other issues:
- dots in file names
- path component contains a keyword

Given that partials and objects don't share the same namespace, it makes sense to split them apart and avoid trying to hack things in.

This also means that `ast::lookup_path` is gone since it's no longer shared. Other than that, no changes required.

Reviewed By: vitaut

Differential Revision: D61764388

fbshipit-source-id: 6cc452aab959dc2457dbf56f1a82d78ba2b84ce2