Preserve whitespace kind for partial application offset
Summary:
Currently we're using `source_manager` + `resolved_location` in parser to figure out the column offset of a standalone partial application.

This is buggy because we don't preserve whitespace correctly. For example, `\t` ends up as `<space>` in multi-line partial offsets. This diff fixes that by encoding the actual whitespace string into the ast.

Reviewed By: yoney

Differential Revision: D61897650

fbshipit-source-id: 3f54838a69269dd6fd3c9e23f376234cab88b6f2