Integrating AnyPatch into DynamicPatch
Summary:
This will also change DynamicUnknownPatch -- now if we have a patch that assigns struct, we won't know whether it's a struct patch or an any patch.

While we can check the field in the assigned struct, this makes it harder to add more fields into Any struct in the future, thus it's better to just mark those assign patch as unknown patch.

Reviewed By: thedavekwon

Differential Revision: D62036816

fbshipit-source-id: 128890068c809a5e98d2759858e0d93dc1688213