a program to parse input with simple-json-protocol
Summary:
This program parses input using the `SimpleJSONProtocolReader` implementation of simple-json-protocol, and then does nothing with the input. With good input, the program just exits success. With bad input, parsing would throw an exception and the program would abort from the unhandled exception.

Can be used with `perf stat` or `perf record` to measure overall performance on a given sample serialization.

Reviewed By: thedavekwon

Differential Revision: D61990602

fbshipit-source-id: 2072512cd0282d58343f3190d0e2b0d1a4bfb5ec