FIx default distribution config
Summary:
Cosco is seeing null pointer exception in SwiftStatsHandler which is most likely due SUM/MAX types not being tracked in the default distribution config for single and multiwindow distributions. This diff adds all Quantile types to default configuration to ensure that blindly looking up a valid quantile in the map should be safe. The alternative way to handle this gracefully is to elect a desired distrobution config and export stats like the following:

```
distributions.forEach(
        (key, dist) -> {
          addQuantilesToCounters(key, ONE_MINUTE, resultCounters, dist.getOneMinuteQuantiles());
          addQuantilesToCounters(key, ONE_HOUR, resultCounters, dist.getOneHourQuantiles());
          addQuantilesToCounters(key, "", resultCounters, dist.getAllTimeQuantiles());
        });

 private void addQuantilesToCounters(
      String baseKey, String windowKey, Map<String, Long> counters, Map<Quantile, Long> quantiles) {
    for (Map.Entry<Quantile, Long> entry : quantiles.entrySet()) {
      counters.put(baseKey + "." + entry.getKey().getKey() + windowKey, entry.getValue());
    }
```

This allow configured quantiles to drive name creation in the final counter map.

Reviewed By: sangongs

Differential Revision: D62077597

fbshipit-source-id: 46edc8a92476b43fa671097d642fb0e0e81679f2