Remove unnecessary token check in parse_partial_lookup
Summary:
`tok::dot` cannot appear in partial lookups because `tok::path_component` consumes the `.` character in `lexer::state_partial_application`. Furthermore, looking just above, we see that `.` is not listed as an allowed token in the grammar. This is a no-op in practice.

Also changing variable names since path components are not identifiers.

Reviewed By: yoney

Differential Revision: D62206922

fbshipit-source-id: fe4c6cd646fa303ac68e5a60853d62028dbf9971