Introduce whisker::expected
Summary:
This is cherry-picking the "bang for your buck" parts of `std::expected` so that we can have a vocabulary type NOW that we can replace with `std::expected` in C++23.

At this point, `whisker::expected` is a subset of the `std::expected` API. We can iteratively converge onto the actual standardized API as needed.

Reviewed By: vitaut

Differential Revision: D62281499

fbshipit-source-id: f7d81601b4e9d532ab75599686c001d705043b68