Create SerializedRequest with empty IOBuf instead of nullptr
Summary:
When we received a request, we call `deserializeMessageBegin` to extract some metadata.

After some checks, we try to trim the remaining bytes before cloning the corresponding buffer in to a `SerializedRequest`. If it was all metadata (ie. no bytes remaining in the buffer), the call to `move()` would return `nullptr`. Change the code to use an empty `IOBuf` instead.

Reviewed By: avalonalex

Differential Revision: D62887479

fbshipit-source-id: 57585a1c22655d74eb61e7e771da208ac9027bad