fix data-race in PayloadSerializer::getInstance
Summary:
A load was improperly marked as load-relaxed when it should be marked as load-acquire (load-consume would technically be permissible as well).

In this example:
```
std::atomic<object*> ptr{};
void producer() { ptr.store(new object(), std::memory_order_release); }
object* consumer() { return ptr.load(std::memory_order_relaxed); }
```

The caller of `consumer` will see either `nullptr` or some pointer. If it seems some pointer, it is likely to access the object through the pointer. But, because the load of the pointer was load-relaxed, the load of the pointer does not *synchronize-with* the store-release of the pointer. This means that stores which happen-before the store-release of the pointer in one thread are not necessarily visible after the load-relaxed of the same pointer value if the load-relaxed happens in a different thread.

The solution is for `consumer` to use load-acquire. This means that stores which happen-before the store-release of the pointer in one thread are visible after the load-acquire of the same pointer value even if the load-acquire happens in a different-thread.

Reviewed By: robertroeser

Differential Revision: D63836571

fbshipit-source-id: d290ad57b70109122e71a3e7ba6757491cae7ebf