remove intermediate py3 base class
Summary:
Right now, thrift-py3 enum is a subclass of `py3.types.Enum`, which is a subtype of `python.types.Enum`. Having an extra base class in the py3 enum chain increases the cost of method lookups. It also makes it harder to claim that py3 enums are actually equivalent to python enums.

The job of `py3.types.Enum` is to implement `typing.Supports{Int,Index}`, i.e., `__int__` and `__index__`. The reason is that thrift-python enums are direct subclasses of `int` but `py3` is not. The py3 behavior is actually preferable; we'll eventually migrate python to not be a subclass of `int`, and instead just be `Supports{Int,Index}`.

As a bonus, we can switch `self.value` calls to `self._fbthrift_value_`, because the former is actually a `property` method call.

As a next step, need to better align the union discriminant enum to be more like that in `thrift-python`, which uses python standard `enum.Enum`.

V2: buildall

Reviewed By: yoney

Differential Revision: D63903798

fbshipit-source-id: e172af94fd728c78378cacb5c34911f995b58084