Reduce duplication in the generation of typing.Final.
Summary:
Generate `typing.Final` with a check for `generate_immutable_types` and eliminate the significant duplication of mustache text.

Note: This was part of a later change later up in the stack. It was more of a refactor of code that already existed rather than part of that change. Given this change can live on its own and didn't need to pollute that other change, I chose to split it into its own diff.

Reviewed By: yoney

Differential Revision: D64275093

fbshipit-source-id: 4aa0e4c9b1f65860b4110b2296c4b63b8134a235