Add a try catch for event callbacks in thrift python server
Summary:
Event hooks in thrift are allowed to throw, but the thrift python server just lacks a try/catch, casing the exception to escape the server and propagate up, which results in the client's connection getting dropped and the client just timing out waiting for a response.

Adding a try/catch to properly return an ApplicationException to the user.

Reviewed By: ahilger

Differential Revision: D63782630

fbshipit-source-id: 8ce03df7acc64b024863d876d4e75f2d268c4a1e