Use setException in process method
Summary: Uses the setException method of the event handler instead of catching multiple exception types. The setException method will handle 0..n types of exceptions as declared in the .thrift file, in the event of zero declared exceptions, it will always return false and trigger wrapping the exception into an ApplicationException.

Reviewed By: yoney

Differential Revision: D64336980

fbshipit-source-id: c7c1379c40a49a605cefc3e7d03b26980546599e