Reduce binary size by templatizing TEnumTraits::typeName
Summary:
Previously diff D64361901 was reverted due to the increasing binary size.

The problem was that the `type_name` was a data member that was always added to the binary. This diff changed to use template function so that it only gets instantiated when used. Given that typeName is rarely used, this helps reduce binary size.

Differential Revision: D64478031

fbshipit-source-id: 164ebf5366b54c087021bca4c40407d78c0f8222