Add observability for pending connections
Summary:
# Issue
Thrift currently does not have any observability into the number of pending connections which are waiting for acceptance by the IO Worker threads.

This observability is especially important during connection storms, which may possibly overload the service with the large number of incoming connections causing OOMs, which was the case for the uCache service (S399591, S400702)

# Change
Introduce observability into pending connections that are enqueued for the IO workers for processing of heavyweight TLS Handshake and other connection acceptance related operations.

**Counter prefix:** `thrift.pending_connections`

**Counter types:**
- P50 - median,  helps to understand the "average" pending connections number under normal conditions.
- P99 - captures rare but impactful high-congestion events.
- P100 - reports maximum number of pending connections observed.

Example values:
 {F1936343509}

Reviewed By: robertroeser

Differential Revision: D64355577

fbshipit-source-id: 935281b926add87adae4d045befd81b2f5286ffc