Cleanup semantic analysis
Summary:
`schematizer` is part of the semantic analysis (an optional mutation stage), so move it to `sema` where it belongs and put it in the `detail` namespace since it's not supposed to be consumed by users directly. This reduces the public API surface.

Fix includes and naming conventions while at it and move `value_id` to `schematizer` since it's no longer used in AST.

Reviewed By: yoney

Differential Revision: D64605531

fbshipit-source-id: 04c45e2517da98c6110879cc36904d056870fa8f