Test to ensure deterministic serialization order
Summary:
Add a unittest for verifying field serialization order.

Sigh.

I don't think Thrift spec requires a certain order for field serialization.
This [post](https://fb.workplace.com/groups/thriftusers/permalink/725328998176515/) seems to suggest the same.

However, there are users who assume that the order is deterministic. And they use serialized bytes as keys in memcached (for example).

So, we have to keep this behavior going forward. I am adding a unittest to ensure that this behavior is documented and stays the same.

Reviewed By: awalterschulze

Differential Revision: D64727545

fbshipit-source-id: 5148d8eb4edbcee142842d09a70f86d8f6aff90e