When HTTP is detected, log part of the request to help determine the cause.
Summary: In many cases the Thrift server code can detect when a client is incorrectly trying to do HTTP to the server.  When the server has detected an HTTP request this patch has it output the first 48 bytes of the request to the log.  The goal is to capture the URL being requested (e.g. "GET /foo/bar/baz?option") to provide additional assistance in tracking down the errant client.

Reviewed By: awalterschulze

Differential Revision: D64760209

fbshipit-source-id: 60dd056186bbf2bb45f0f595fab8707f4f33877d