Fix potential stack-use-after-free in HandlerCallback
Summary: When we do `startOnExecutor(doProcess(...))`, technically the coroutine implementation is the non-static data member of the lambda type backing `doProcess`. However, `doProcess` goes out of scope. Since we don't have captures, that seems ok in practice but let's take the guesswork out by moving the logic out to a static member.

Reviewed By: avalonalex

Differential Revision: D64267142

fbshipit-source-id: af5df5d614041d2b9cbb6595fe2e2e76ee15fa9b