Fix an issue in thrift py3 where cpp adapted fields didn't get hidden leading to potential compilation issues.
Summary:
The `cpp.adapter` annotation breaks thrift py3 and isn't intended to be usable in that thrift flavour, and we had intended for any cpp adapted field to become hidden in thrift py3, but the logic for hiding fields in thrift py3 incorrectly just checked for `python.py3hidden` annotations.

Fixing this by checking for `cpp.adapter` annotations in the t_named `is_hidden` check and calling that when checking hidden fields.

#buildall

Reviewed By: ahilger

Differential Revision: D64867178

fbshipit-source-id: 3d0bc3d4d40bbb4631a99ebf4966561789b02f28